feat: Stage 1 - Create package structure and extract base utilities
- Created slack_mcp package directory structure
- Extracted utility functions into separate modules:
  - client.py: Slack client initialization and management
  - errors.py: Error handling and rate limiting
  - formatting.py: Data formatting utilities
- Updated slack_mcp_server.py to use the new package structure
- Maintained backward compatibility with import fallback
- Added README.md documenting the refactoring progress

This is the first stage of the modularization effort to improve
code organization and maintainability.
