+++ src/slack_mcp/README.md
@@ -1,0 +1,58 @@
+# Slack MCP Package Structure
+
+This directory contains the modularized Slack MCP Server implementation.
+
+## Structure
+
+```
+slack_mcp/
+├── __init__.py          # Package initialization
+├── utils/               # Utility modules
+│   ├── __init__.py      # Utils package exports
+│   ├── client.py        # Slack client management
+│   ├── errors.py        # Error handling and rate limiting
+│   └── formatting.py    # Message and data formatting
+├── models/              # Pydantic models (Stage 2)
+│   └── __init__.py
+└── tools/               # Tool implementations (Stages 3-7)
+    └── __init__.py
+```
+
+## Stage 1: Package Structure and Base Utilities (Completed)
+
+### What was done:
+1. Created the package directory structure
+2. Extracted utility functions from the monolithic server file:
+   - **client.py**: Slack client initialization and management
+     - `validate_slack_token()`: Validates and initializes Slack client
+     - `init_async_client()`: Initializes async Slack client
+     - `get_slack_client()`, `set_slack_client()`: Client accessors
+     - `MCP_USER_AGENT`: User agent string for API calls
+   
+   - **errors.py**: Error handling and rate limiting
+     - `rate_limit_check()`: Implements rate limiting with delays
+     - `make_slack_request()`: Wrapper for API calls with error handling
+     - `MIN_API_INTERVAL`: Rate limit interval constant
+   
+   - **formatting.py**: Data formatting utilities
+     - `format_file_info()`: Formats file information for display
+
+3. Updated imports in `slack_mcp_server.py` to use the new package structure
+4. Maintained backward compatibility with try/except import fallback
+
+### Key Design Decisions:
+- Used underscore prefix for internal global variables in client.py
+- Exposed public API through getter/setter functions
+- Maintained all original functionality without breaking changes
+- Added proper __all__ exports in __init__.py files
+
+## Next Steps
+
+### Stage 2: Extract Pydantic Models
+Move all Pydantic model definitions to `models/` subdirectory
+
+### Stage 3-7: Extract Tool Groups
+Organize tools by functionality into separate modules under `tools/`
+
+### Stage 8-10: Testing and Documentation
+Add comprehensive tests and update documentation
