+++ src/slack_mcp/utils/__init__.py
@@ -1,0 +1,37 @@
+"""Utility modules for Slack MCP Server."""
+
+# Client utilities
+from .client import (
+    MCP_USER_AGENT,
+    validate_slack_token,
+    init_async_client,
+    get_slack_client,
+    set_slack_client,
+    get_async_slack_client,
+)
+
+# Formatting utilities
+from .formatting import format_file_info
+
+# Error handling and rate limiting utilities
+from .errors import (
+    MIN_API_INTERVAL,
+    rate_limit_check,
+    make_slack_request,
+)
+
+__all__ = [
+    # Client utilities
+    "MCP_USER_AGENT",
+    "validate_slack_token",
+    "init_async_client",
+    "get_slack_client",
+    "set_slack_client",
+    "get_async_slack_client",
+    # Formatting utilities
+    "format_file_info",
+    # Error handling utilities
+    "MIN_API_INTERVAL",
+    "rate_limit_check",
+    "make_slack_request",
+]
